strftime() needs to be locked.
On some systems, it just needs to not have the locale nor environment
change while it is executing.  A many-reader lock for each will work for
those.  The Posix standard, on the other hand, says (under the tzset
entry) that strftime effective calls it, thus setting some globals, so
it needs exclusive access for those.  A Configure probe could be written
to check for the behavior.