Time-Piece/Piece.xs: Use isDIGIT, not isdigit
isDIGIT is not locale-dependent, matching just [0-9].  There was a bug
using isdigit() in at least one place where it calculated the value of
the digit by subtracting the original of '0' from it.  This works only
if the digit is in [0-9], but there are locales where there is another
set of digits, and so would give a very wrong result.  That makes it
clear to me that only [0-9] were assumed by the original authors.